Provide a good commit message example
The purpose of this commit is to provide an example of a band-crafted,
artisanala commit message. The first line is a short, approximately 50 character
summary, followed by an empty line. The subsequent paragraphs are jam-packed
with descriptive information about the chanmge, but each line is kept under 72
charcaters in length.

If even more information is needed to explain the change, more paragraphs can be added afeter blank lines, with links to isssues, tockets, or bugs, Remember
that future you will thank current you for your thoughtfulness and foresight!